# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog Berita API",
    "version": "1.0.0",
    "description": "API documentation for Blog Berita application. This API provides endpoints for managing news articles, categories, tags, comments, advertisements, and user authentication."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://backend-blog-berita.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/api/posts/{slug}/view": {
      "post": {
        "tags": [
          "Blog Public"
        ],
        "summary": "Record page view",
        "description": "Record a page view for a post. Requires X-Visitor-Id header for unique visitor tracking.",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Post slug"
          },
          {
            "in": "header",
            "name": "X-Visitor-Id",
            "schema": {
              "type": "string"
            },
            "description": "Unique visitor identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "View recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "View recorded"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/admin/tags/{id}": {
      "put": {
        "tags": [
          "Admin Tags"
        ],
        "summary": "Update a tag",
        "description": "Update an existing tag by ID (requires admin)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Tag ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "slug"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "JS"
                  },
                  "slug": {
                    "type": "string",
                    "example": "js"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin only"
          },
          "404": {
            "description": "Tag not found"
          },
          "409": {
            "description": "Slug already used"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin Tags"
        ],
        "summary": "Delete a tag",
        "description": "Delete a tag by ID (requires admin)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Tag ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Tag deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin only"
          },
          "404": {
            "description": "Tag not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/admin/comments": {
      "get": {
        "tags": [
          "Admin Comments"
        ],
        "summary": "Get all comments (paginated)",
        "description": "Retrieve all comments with pagination (requires admin)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated comments list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "total": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin only"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/admin/categories/{id}": {
      "put": {
        "tags": [
          "Admin Categories"
        ],
        "summary": "Update a category",
        "description": "Update an existing category by ID (requires admin)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Category ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "slug"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Teknologi Update"
                  },
                  "slug": {
                    "type": "string",
                    "example": "teknologi-update"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin only"
          },
          "404": {
            "description": "Category not found"
          },
          "409": {
            "description": "Slug already used"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin Categories"
        ],
        "summary": "Delete a category",
        "description": "Delete a category by ID (requires admin)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Category ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin only"
          },
          "404": {
            "description": "Category not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}